Allow `const_assert_{eq,ne}!` to take 2+ arguments
They only took 2 arguments so that error messages can be added later
when `if` in `const` becomes stable. However, adding that will be a
breaking change due to a bump in minimum supported Rust version. So this
change is fine to have in the meantime, and then removed later once that
becomes how this macro is implemented.