[Cargo] Update badges
Removed Travis CI badge because GitHub Actions are used instead.

Removed "is it maintained" badges because they're not very indicative of
maintenance for this project, since issues are used to track features,
and less so bugs.

Changed "maintenance" badge to "actively maintained" since it's more
indicative of this project's status.